backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
alembic>=1.14.0
celery[redis]>=5.4.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
pyjwt>=2.10.0
bcrypt>=4.2.0
cryptography>=44.0.0
ansible-core>=2.17.0
ansible-runner>=2.4.0
python-multipart>=0.0.18
redis>=5.2.0
pywinrm>=0.4.3
httpx>=0.28.0
python-nmap>=0.7.1
pytest>=8.3.0
pytest-asyncio>=0.24.0
